removed default media player option + cleanup
