+++ backend/pkg/settings/generator_test.go
@@ -481,3 +481,15 @@ func TestGenerateConfigYaml_EdgeCases(t *testing.T) {
 		})
 	}
 }
+
+func TestGenerateConfigYaml_defaultPermissionsDocumentsView(t *testing.T) {
+	cfg := SetDefaults(true)
+	cfg.Server.Sources = []*Source{{Path: ".", Config: SourceConfig{}}}
+	out, err := GenerateConfigYamlWithSource(&cfg, true, true, true, ".")
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !strings.Contains(out, "defaultPermissions:") || !strings.Contains(out, "view:") {
+		t.Fatalf("expected defaultPermissions to include view, got excerpt around block")
+	}
+}
