+++ backend/swagger/docs/swagger.yaml
@@ -1039,9 +1039,6 @@ definitions:
       debugOffice:
         description: debug onlyoffice editor
         type: boolean
-      defaultMediaPlayer:
-        description: disable the styled feature-rich media player for browser default
-        type: boolean
       disableOnlyOfficeExt:
         description: list of file extensions to disable onlyoffice editor for
         type: string
@@ -1763,9 +1760,6 @@ definitions:
         type: boolean
       autoplayMedia:
         description: autoplay media files in preview
-        type: boolean
-      defaultMediaPlayer:
-        description: disable html5 media player and use the default media player
         type: boolean
       disableHideSidebar:
         description: disable the hide sidebar preview for previews and editors
