+++ frontend/src/store/types.ts
@@ -78,7 +78,6 @@ export interface UserObject {
     models: boolean;
     popup: boolean;
     autoplayMedia?: boolean;
-    defaultMediaPlayer?: boolean;
   };
   loginType: string;
   username: string;
